Issue #11163: Enforce file size on InputLocalVariableName
Issue #11163: added import to methods
